inventory low stock query
not currently working

+++ database/migrations/initial_database.php
@@ -240,7 +240,7 @@ return new class extends Migration
 
         Schema::create('store_item_storage', function (Blueprint $table) {
             $table->foreignID('store_item_id')->references('id')->on('store_item')->onDelete('cascade');
-            $table->double('quantity');
+            $table->integer('quantity');
             $table->foreignID('location_id')->references('id')->on('location')->onDelete('cascade');
             $table->timestamp('expiration_date')->nullable();
         });
